refactor: Re-organise the IO related namespaces (#137)
- Move IO one namespace up (out of the `Input` namespace)
- Move all the output related classes to a `Output` namespace
