+++ doc/call-other-commands.md
@@ -24,7 +24,7 @@ namespace App\Console\Command;
 use Fidry\Console\Command\CommandAware;
 use Fidry\Console\Command\CommandAwareness;
 use Fidry\Console\Command\Command;
-use Fidry\Console\Input\IO;
+use Fidry\Console\IO;
 use Symfony\Component\Console\Input\ArrayInput;
 
 final class CreateUserCommand implements Command, CommandAware
