+++ src/Command/Command.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace Fidry\Console\Command;
 
 use Fidry\Console\ExitCode;
-use Fidry\Console\Input\IO;
+use Fidry\Console\IO;
 
 interface Command
 {
