+++ src/Output/Compatibility/StyledOutputSymfony6.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Fidry\Console\Input\Compatibility;
+namespace Fidry\Console\Output\Compatibility;
 
 use Symfony\Component\Console\Helper\ProgressBar;
 use Symfony\Component\Console\Helper\Table;
