+++ src/Output/DecoratesOutput.php
@@ -11,25 +11,21 @@
 
 declare(strict_types=1);
 
-namespace Fidry\Console\Input;
+namespace Fidry\Console\Output;
 
 use Composer\InstalledVersions;
-use Fidry\Console\Input\Compatibility\StyledOutputSymfony5;
-use Fidry\Console\Input\Compatibility\StyledOutputSymfony6;
+use Fidry\Console\Output\Compatibility\DecoratesOutputSymfony5;
+use Fidry\Console\Output\Compatibility\DecoratesOutputSymfony6;
 use function Safe\class_alias;
 use function version_compare;
 
-// This is purely for the compatibility layer between Symfony5 & Symfony6. The
-// behaviour is the same, only the method signatures differ.
-// To have a more comprehensive look of the class check:
-// stubs/StyledOutput.php
 class_alias(
     (string) version_compare(
         (string) InstalledVersions::getPrettyVersion('symfony/console'),
         'v6.0',
         '>=',
     )
-        ? StyledOutputSymfony6::class
-        : StyledOutputSymfony5::class,
-    \Fidry\Console\Input\StyledOutput::class,
+        ? DecoratesOutputSymfony6::class
+        : DecoratesOutputSymfony5::class,
+    \Fidry\Console\Output\DecoratesOutput::class,
 );
