+++ src/Output/SymfonyStyledOutput.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Fidry\Console\Input;
+namespace Fidry\Console\Output;
 
 use Symfony\Component\Console\Style\SymfonyStyle;
