+++ tests/Application/Fixture/FailingCommand.php
@@ -15,7 +15,7 @@ namespace Fidry\Console\Tests\Application\Fixture;
 
 use Fidry\Console\Command\Command;
 use Fidry\Console\Command\Configuration;
-use Fidry\Console\Input\IO;
+use Fidry\Console\IO;
 use UnexpectedValueException;
 
 final class FailingCommand implements Command
